Rename tools -> uxy; move uxy script -> uxy.__main__; add pyproject.toml
