+++ pyproject.toml
@@ -1,0 +1,22 @@
+[tool.poetry]
+name = "uxy"
+version = "0.1.0"
+description = "Adding structure to the UNIX tools"
+authors = ["sustrik <sustrik@250bpm.com>"]
+license = "MIT"
+readme = "README.md"
+homepage = "https://github.com/sustrik/uxy"
+repository = "https://github.com/sustrik/uxy.git"
+
+[tool.poetry.scripts]
+uxy = "uxy.__main__:main"
+
+[tool.poetry.dependencies]
+python = "^3.6"
+PyYAML  = "4.2b1"
+
+[tool.poetry.dev-dependencies]
+
+[build-system]
+requires = ["poetry>=0.12"]
+build-backend = "poetry.masonry.api"
