+++ uxy/uxy_from_csv.py
@@ -17,29 +17,28 @@
 #  IN THE SOFTWARE.
 
 import argparse
-import json
+import io
+import csv
 
-from tools import base
+from uxy import base
 
-def to_json(args, uxy_args):
+def from_csv(args, uxy_args):
   parser = argparse.ArgumentParser()
-  subp = parser.add_subparsers().add_parser('to-json',
-    help="convert UXY to JSON")
+  subp = parser.add_subparsers().add_parser('from-csv',
+    help="convert CSV to UXY")
   args = parser.parse_args(args)
 
-  s = base.stdin.readline()
-  hdr = base.split_fields(s)
-  base.writeline("[\n")
-  first = True
+  # Read the headers
+  ln = base.stdin.readline()
+  r = csv.reader(io.StringIO(ln))
+  for fields in r:
+    fields = " ".join([base.encode_field(f) for f in fields])
+    fmt = base.Format(fields)
+    base.writeline(fields + "\n")
   for ln in base.stdin:
-    if not first:
-      base.writeline(",\n")
-    else:
-      first = False
-    fields = base.split_fields(ln)
-    item = {}
-    for i in range(0, len(fields)):
-      item[base.decode_field(hdr[i])] = base.decode_field(fields[i])
-    base.writeline(json.dumps(item, indent=4))
-  base.writeline("\n]\n")
+    r = csv.reader(io.StringIO(ln))
+    for fields in r:
+      fields = [base.encode_field(f) for f in fields]
+      base.writeline(fmt.render(fields))
+
   return 0
